chore(native): refactor span module
