+++ src/native/span/span_event.rs
@@ -1,0 +1,29 @@
+use pyo3::{
+    types::{PyDict, PyTuple},
+    Bound, PyAny, PyResult,
+};
+
+#[pyo3::pyclass(name = "SpanEventData", module = "ddtrace.internal._native", subclass)]
+#[derive(Default)]
+pub struct SpanEventData {}
+
+#[pyo3::pymethods]
+impl SpanEventData {
+    #[new]
+    #[pyo3(signature =(
+        *_py_args,
+        **_py_kwargs,
+    ))]
+    pub fn __new__(_py_args: &Bound<'_, PyTuple>, _py_kwargs: Option<&Bound<'_, PyDict>>) -> Self {
+        Self::default()
+    }
+
+    pub fn __init__(
+        &mut self,
+        _name: &Bound<'_, PyAny>,
+        _attributes: Option<&Bound<'_, PyAny>>,
+        _time_unix_nano: Option<u64>,
+    ) -> PyResult<()> {
+        Ok(())
+    }
+}
