+++ src/native/span/span_link.rs
@@ -1,0 +1,41 @@
+use pyo3::{
+    types::{PyDict, PyInt, PyTuple},
+    Bound, PyAny, PyResult,
+};
+
+#[pyo3::pyclass(name = "SpanLinkData", module = "ddtrace.internal._native", subclass)]
+#[derive(Default)]
+pub struct SpanLinkData {}
+
+#[pyo3::pymethods]
+impl SpanLinkData {
+    #[new]
+    #[pyo3(signature =(
+        *_py_args,
+        **_py_kwargs,
+    ))]
+    pub fn __new__(_py_args: &Bound<'_, PyTuple>, _py_kwargs: Option<&Bound<'_, PyDict>>) -> Self {
+        Self::default()
+    }
+
+    #[pyo3(signature = (
+        trace_id,
+        span_id,
+        tracestate = None,
+        flags = None,
+        attributes = None,
+        _dropped_attributes = 0,
+    ))]
+    #[allow(unused_variables)]
+    pub fn __init__<'p>(
+        &mut self,
+        trace_id: &Bound<'p, PyInt>,
+        span_id: &Bound<'p, PyInt>,
+        tracestate: Option<&Bound<'p, PyAny>>,
+        flags: Option<&Bound<'p, PyInt>>,
+        attributes: Option<&Bound<'p, PyAny>>,
+        _dropped_attributes: u32,
+    ) -> PyResult<()> {
+        Ok(())
+    }
+}
